Revert "cpufreq: schedutil: Don't restrict kthread to related_cpus unnecessarily"
This reverts commit e2cabe48c20efb174ce0c01190f8b9c5f3ea1d13.

We observed on Hikey960 cases where the sugov threads for both clusters
were running on the big CPUs. With EAS, the big cluster can stay idle
for longs period of time if there is not a lot of activity in the
system. In this case, the sugov thread for the little side can wake-up
often and prevent the big cluster to go deeply idle, hence costing a lot
of energy.

Since the benefits from not restricting the CPUs on which the kthreads
can run is less than clear, fallback on the old behaviour which works
just fine.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>