stop_machine: Reflow cpu_stop_queue_two_works()
The code flow in cpu_stop_queue_two_works() is a little arcane; fix
this by lifting the preempt_disable() to the top to create more natural
nesting wrt the spinlocks and make the wake_up_q() and preempt_enable()
unconditional at the end.

Furthermore, enable preemption in the -EDEADLK case, such that we
spin-wait with preemption enabled.

Cc: isaacm@codeaurora.org
Cc: pkondeti@codeaurora.org
Cc: mingo@kernel.org
Cc: psodagud@codeaurora.org
Cc: matt@codeblueprint.co.uk
Cc: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Cc: gregkh@linuxfoundation.org
Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: http://lkml.kernel.org/r/20180730112140.GH2494@hirez.programming.kicks-ass.net